fix: update test script to most recent version
