+++ package.json
@@ -6,8 +6,9 @@
     "build": "scripts/build.sh deploy",
     "start": "run-p ganache deploy",
     "start:full": "run-p ganache migrate:full",
-    "coverage": "export COVERALLS_REPO_TOKEN='qiXNelh1HXMAsoPah9yRD2yG2WlqrDfLh' && run-s test:generate-typings && run-s test:tsc && scripts/coverage.sh",
+    "coverage": "export COVERALLS_REPO_TOKEN='qiXNelh1HXMAsoPah9yRD2yG2WlqrDfLh' && run-s test:generate-typings && run-s test:tsc && run-s coveragescript",
     "testscript": "scripts/test.sh",
+    "coveragescript": "scripts/coverage.sh",
     "ganache": "ganache-cli -m \"$MNEMONIC\" -g 1000000 -l 8000000 --port 8545 --networkId 4447 -q",
     "wait": "wait-on tcp:9545",
     "minimize": "scripts/minimize.sh 'build/contracts'",
