+++ scripts/coverage.sh
@@ -1,3 +1,3 @@
 #!/usr/bin/env bash
 
-SOLIDITY_COVERAGE=true CONTINUOUS_INTEGRATION=true scripts/test.sh
+SOLIDITY_COVERAGE=true CONTINUOUS_INTEGRATION=true scripts/test.sh "$@"
