+++ test/RelayingReserve.integration.test.ts
@@ -1,4 +1,4 @@
-import * as helpers from '../helpers';
+import * as helpers from './helpers';
 
 const Identity = artifacts.require("Identity");
 const DaoCreatorGoodDollar = artifacts.require("DaoCreatorGoodDollar");
@@ -7,16 +7,16 @@ const GoodDollar = artifacts.require("GoodDollar");
 const ControllerInterface = artifacts.require("ControllerInterface");
 const AbsoluteVote = artifacts.require("AbsoluteVote");
 const SchemeRegistrar = artifacts.require("SchemeRegistrar");
-const ReserveMinter = artifacts.require("ReserveMinter");
+const ReserveRelayer = artifacts.require("ReserveRelayer");
 
-contract("ReserveMinter - Minting to reserve", ([founder, whitelisted, receiver]) => {
+contract("ReserveRelayer - Transferring reserve", ([founder, whitelisted, receiver]) => {
 
 	let identity: helpers.ThenArg<ReturnType<typeof Identity['new']>>;
 	let avatar: helpers.ThenArg<ReturnType<typeof Avatar['new']>>;
 	let controller: helpers.ThenArg<ReturnType<typeof ControllerInterface['new']>>;
 	let absoluteVote: helpers.ThenArg<ReturnType<typeof AbsoluteVote['new']>>;
 	let token: helpers.ThenArg<ReturnType<typeof GoodDollar['new']>>;
-	let reserveMinter: helpers.ThenArg<ReturnType<typeof ReserveMinter['new']>>;
+	let reserveRelayer: helpers.ThenArg<ReturnType<typeof ReserveRelayer['new']>>;
 
 	let proposalId: string;
 
@@ -31,41 +31,59 @@ contract("ReserveMinter - Minting to reserve", ([founder, whitelisted, receiver]
 		controller = await ControllerInterface.at(await avatar.owner());
 		absoluteVote = await AbsoluteVote.deployed();
 		token = await GoodDollar.at(await avatar.nativeToken());
-		reserveMinter = await ReserveMinter.new(avatar.address, helpers.toGD("10"), receiver);
+		reserveRelayer = await ReserveRelayer.new(avatar.address, identity.address, receiver, periodStart, periodEnd);
 		await identity.addWhitelisted(whitelisted);
 	});
 
 	it("should not allow relayer with null address receiver", async () => {
 		const periodStart = (await web3.eth.getBlock('latest')).timestamp + periodOffset;
 		const periodEnd = periodStart + periodOffset;
-		helpers.assertVMException(ReserveMinter.new(avatar.address, helpers.toGD("10"), helpers.NULL_ADDRESS), "receiver cannot be null address");
+		helpers.assertVMException(ReserveRelayer.new(avatar.address, identity.address, helpers.NULL_ADDRESS, periodStart, periodEnd), "receiver cannot be null address");
 	});
 
-	it("should correctly propose ReserveMinter scheme", async () => {
+	it("should perform transactions and increase fee reserve", async () => {
+	  const oldReserve = await token.balanceOf(avatar.address);
+
+	  await token.transfer(whitelisted, helpers.toGD("10"));
+	  await token.transfer(whitelisted, helpers.toGD("10"));
+	  await token.transfer(whitelisted, helpers.toGD("10"));
+
+	  // Check that reserve has received fees
+	  const reserve = (await token.balanceOf(avatar.address)) as any;
+
+	  const reserveDiff = reserve.sub(oldReserve);
+	  const totalFees = ((await token.getFees(helpers.toGD("10"))) as any).mul(new (web3 as any).utils.BN("3"));
+	  expect(reserveDiff.toString()).to.be.equal(totalFees.toString());
+	});
+
+	it("should correctly propose ReserveRelayer scheme", async () => {
 		const schemeRegistrar = await SchemeRegistrar.deployed();
-		const transaction = await schemeRegistrar.proposeScheme(avatar.address, reserveMinter.address,
+		const transaction = await schemeRegistrar.proposeScheme(avatar.address, reserveRelayer.address,
 			helpers.NULL_HASH, "0x00000010", helpers.NULL_HASH);
 
 		proposalId = transaction.logs[0].args._proposalId;
 	});
 
-	it("should correctly register ReserveMinter scheme", async () => {
+	it("should correctly register ReserveRelayer scheme", async () => {
 		const voteResult = await absoluteVote.vote(proposalId, 1, 0, founder);
 		const executeProposalEventExists = voteResult.logs.some(e => e.event === 'ExecuteProposal');
 
 		assert(executeProposalEventExists);
 	});
 
-	it("should start, mint to receiver and then end", async () => {
+	it("should start, transfer reserve and then end", async () => {
+		await helpers.assertVMException(reserveRelayer.start(), "not in period");
+		await helpers.increaseTime(periodOffset*1.5);
+
 		const oldBalance = await token.balanceOf(receiver);
 		expect(oldBalance.toString()).to.be.equal("0");
 
 		const reserve = (await token.balanceOf(avatar.address)) as any;
 		
-		assert(await reserveMinter.start());
+		assert(await reserveRelayer.start());
 
 		const newBalance = await token.balanceOf(receiver);
 
-		expect(newBalance.toString()).to.be.equal((helpers.toGD("10")).toString());
+		expect(newBalance.toString()).to.be.equal(reserve.toString());
 	});
 })
